renamed widgets and reorganised files
